requirements.txt
unstructured
python-dotenv~=1.1.0
langchain_aws~=0.2.23
langchain_core~=0.3.60
langchain_community~=0.3.24
langchain_text_splitters~=0.3.8
boto3~=1.38.19
langgraph~=0.4.5
nltk
libmagic
pyyaml~=6.0.2
langchain_neo4j
openai
pyjwt~=2.10.1
requests~=2.32.3
pandas~=2.2.3
numpy~=1.26.4
typing_extensions~=4.13.2
uvicorn~=0.34.2
fastapi~=0.115.12
pydantic~=2.11.4
pyjwt